Things to their right place
